+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 # 0.2.0
+- Explicitly sort sidepanel images/containers by newest